+++ tests/neurst_pt/decoders/transformer_decoder_test.py
@@ -79,49 +79,49 @@ def test_transformer_decoder_prenorm():
     pt_decoder._output_norm_layer.weight.data = torch.FloatTensor(tf_decoder._output_norm_layer.gamma.numpy())
     pt_decoder._output_norm_layer.bias.data = torch.FloatTensor(tf_decoder._output_norm_layer.beta.numpy())
     pt_decoder._stacking_layers[0][0]._layer._qkv_transform_layer._kernel.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][0]._layer._qkv_transform_layer._kernel.numpy())
+        tf_decoder._stacking_layers[0]._selfatt_layer._layer._qkv_transform_layer._kernel.numpy())
     pt_decoder._stacking_layers[0][0]._layer._qkv_transform_layer._bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][0]._layer._qkv_transform_layer._bias.numpy())
+        tf_decoder._stacking_layers[0]._selfatt_layer._layer._qkv_transform_layer._bias.numpy())
     pt_decoder._stacking_layers[0][0]._layer._output_transform_layer._kernel.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][0]._layer._output_transform_layer._kernel.numpy())
+        tf_decoder._stacking_layers[0]._selfatt_layer._layer._output_transform_layer._kernel.numpy())
     pt_decoder._stacking_layers[0][0]._layer._output_transform_layer._bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][0]._layer._output_transform_layer._bias.numpy())
+        tf_decoder._stacking_layers[0]._selfatt_layer._layer._output_transform_layer._bias.numpy())
     pt_decoder._stacking_layers[0][0]._norm_layer.weight.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][0]._norm_layer.gamma.numpy())
+        tf_decoder._stacking_layers[0]._selfatt_layer._norm_layer.gamma.numpy())
     pt_decoder._stacking_layers[0][0]._norm_layer.bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][0]._norm_layer.beta.numpy())
+        tf_decoder._stacking_layers[0]._selfatt_layer._norm_layer.beta.numpy())
 
     pt_decoder._stacking_layers[0][1]._layer._q_transform_layer._kernel.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._layer._q_transform_layer._kernel.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._layer._q_transform_layer._kernel.numpy())
     pt_decoder._stacking_layers[0][1]._layer._q_transform_layer._bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._layer._q_transform_layer._bias.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._layer._q_transform_layer._bias.numpy())
     pt_decoder._stacking_layers[0][1]._layer._kv_transform_layer._kernel.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._layer._kv_transform_layer._kernel.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._layer._kv_transform_layer._kernel.numpy())
     pt_decoder._stacking_layers[0][1]._layer._kv_transform_layer._bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._layer._kv_transform_layer._bias.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._layer._kv_transform_layer._bias.numpy())
     pt_decoder._stacking_layers[0][1]._layer._output_transform_layer._kernel.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._layer._output_transform_layer._kernel.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._layer._output_transform_layer._kernel.numpy())
     pt_decoder._stacking_layers[0][1]._layer._output_transform_layer._bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._layer._output_transform_layer._bias.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._layer._output_transform_layer._bias.numpy())
     pt_decoder._stacking_layers[0][1]._norm_layer.weight.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._norm_layer.gamma.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._norm_layer.gamma.numpy())
     pt_decoder._stacking_layers[0][1]._norm_layer.bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][1]._norm_layer.beta.numpy())
+        tf_decoder._stacking_layers[0]._crossatt_layer._norm_layer.beta.numpy())
 
     pt_decoder._stacking_layers[0][2]._layer._dense1.weight.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][2]._layer._conv1.kernel.numpy().transpose([1, 0]))
+        tf_decoder._stacking_layers[0]._ffn_layer._layer._conv1.kernel.numpy().transpose([1, 0]))
     pt_decoder._stacking_layers[0][2]._layer._dense1.bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][2]._layer._conv1.bias.numpy())
+        tf_decoder._stacking_layers[0]._ffn_layer._layer._conv1.bias.numpy())
     pt_decoder._stacking_layers[0][2]._layer._dense2.weight.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][2]._layer._conv2.kernel.numpy().transpose([1, 0]))
+        tf_decoder._stacking_layers[0]._ffn_layer._layer._conv2.kernel.numpy().transpose([1, 0]))
     pt_decoder._stacking_layers[0][2]._layer._dense2.bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][2]._layer._conv2.bias.numpy())
+        tf_decoder._stacking_layers[0]._ffn_layer._layer._conv2.bias.numpy())
     pt_decoder._stacking_layers[0][2]._norm_layer.weight.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][2]._norm_layer.gamma.numpy())
+        tf_decoder._stacking_layers[0]._ffn_layer._norm_layer.gamma.numpy())
     pt_decoder._stacking_layers[0][2]._norm_layer.bias.data = torch.FloatTensor(
-        tf_decoder._stacking_layers[0][2]._norm_layer.beta.numpy())
+        tf_decoder._stacking_layers[0]._ffn_layer._norm_layer.beta.numpy())
     assert_equal_numpy(tf_decoder(tf_decinp, tf_cache, is_training=False).numpy(),
-                       pt_decoder(pt_decinp, pt_cache, is_training=False).detach().numpy())
+                       pt_decoder(pt_decinp, pt_cache, is_training=False).detach().numpy(), 5e-5)
 
     # for inference
     tf_cache = tf_decoder.create_decoding_internal_cache(tf_inp, tf_inppad, is_inference=True)